# src/openapi.yaml
openapi: 3.0.0
info:
  title: Identity Service
  version: 1.0.0
  description: The Identity Service is responsible for managing user-related operations and system configurations.
    User Management - Create, retrieve, update and delete user accounts.
    Authentication & Authorization - Handle user authentication, session management, and access control.
    System Configurations - Manage system-specific setting such as API configurations, user scopes and groups.
servers:
  - url: /identity-svc/api/v1.0
    description: Local development servers
components:
  parameters:
...
    UserIdParam:
      name: userId
      description: The ID of user to verify
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-F0-9]{8}:[A-F0-9]{4}-[A-F0-9]{4}-[A-F0-9]{4}-[A-F0-9]{11}'
    TokenParam:
      name: token
      description: The token of user to verify
      in: path
      required: true
      schema:
        type: string
  schemas:
    Username:
      type: string
      description: Username for the account
      pattern: '^[a-zA-Z0-9]{4,24}$'
    EmailId:
      type: string
      description: Email id of user
      pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}$'
    Password:
      type: string
      description: user account password
      pattern: '^[a-zA-Z0-9@_.&*#$]{8,32}$'
...
    registerUserPayload:
      type: object
      properties:
        first_name:
          type: string
          description: First Name
          minLength: 1
        last_name:
          type: string
          description: Last Name
        username:
          type: string
          description: Username
          $ref: '#/components/schemas/Username'
        email_id:
          type: string
          description: Email ID
          $ref: '#/components/schemas/EmailId'
        password:
          type: string
          description: Password
          $ref: '#/components/schemas/Password'
      required:
        - first_name
        - username
        - email_id
        - password
    loginUserPayload:
      type: object
      properties:
        userIdentifier:
          type: string
          description: Username or Email Id
          minLength: 1
        password:
          type: string
          description: Password
          $ref: '#/components/schemas/Password'
      required:
        - userIdentifier
        - password
...
    getUserResponseData:
      title: Schema for User Response Data
      type: object
      properties:
        id:
          type: string
        role_id:
          type: string
        role_cd:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        username:
          type: string
        email_id:
          type: string
        login_type:
          type: string
        is_verified:
          type: boolean
        created_date:
          type: string
        modified_date:
          type: string
        login_count:
          type: integer
        last_login:
          type: string
          nullable: true
      required:
        - id
        - role_id
        - role_cd
        - first_name
        - last_name
        - username
        - email_id
        - is_verified
    userInfoResponseData:
      title: Schema for Login User Response Data
      type: object
      properties:
        id:
          type: string
        role_id:
          type: string
        role_code:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        username:
          type: string
        email_id:
          type: string
        login_type:
          type: string
        is_verified:
          type: boolean
        created_date:
          type: string
        modified_date:
          type: string
        login_count:
          type: integer
        last_login:
          type: string
        is_deleted:
          type: boolean
        scopes:
          type: array
        access_token:
          type: string
        refresh_token:
          type: string
        gender:
          type: string
          nullable: true
        dob:
          type: string
          nullable: true
        contact:
          type: string
          nullable: true
        profile_img:
          type: string
          nullable: true
      required:
        - id
        - role_id
        - role_code
        - first_name
        - last_name
        - username
        - email_id
        - login_type
        - is_verified
        - created_date
        - modified_date
        - login_count
        - last_login
        - is_deleted
...
    registerUserResponse:
      type: object
      properties:
        status:
          type: integer
          example: 201
        type:
          type: string
          example: 'SUCCESS'
        message:
          type: string
          example: 'Success'
        devMessage:
          type: string
          example: 'Success'
        success:
          type: boolean
          example: true
        data:
          type: object
          $ref: '#/components/schemas/getUserResponseData'
      required:
        - status
        - type
        - message
        - devMessage
        - success
        - data
    verifyUserResponse:
      type: object
      properties:
        status:
          type: integer
          example: 200
        type:
          type: string
          example: 'SUCCESS'
        message:
          type: string
          example: 'Success'
        devMessage:
          type: string
          example: 'Success'
        success:
          type: boolean
          example: true
        data:
          type: object
      required:
        - status
        - type
        - message
        - devMessage
        - success
        - data
    loginUserResponse:
      type: object
      properties:
        status:
          type: integer
          example: 200
        type:
          type: string
          example: 'SUCCESS'
        message:
          type: string
          example: 'Success'
        devMessage:
          type: string
          example: 'Success'
        success:
          type: boolean
          example: true
        data:
          type: object
          $ref: '#/components/schemas/userInfoResponseData'
      required:
        - status
        - type
        - message
        - devMessage
        - success
        - data
...
    badRequestResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        type:
          type: string
          example: 'BAD_REQUEST'
        message:
          type: string
          example: 'Bad Request'
        devMessage:
          type: string
          example: Bad Request. Required resource does not exist.
        error:
          type: array
        data:
          type: array
...
    notFoundResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        type:
          type: string
          example: 'NOT_FOUND'
        message:
          type: string
          example: 'Resource Not Found'
        devMessage:
          type: string
          example: Not Found. Requested resource does not exist.
        error:
          type: array
        data:
          type: array
    goneResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 410
        type:
          type: string
          example: 'GONE'
        message:
          type: string
          example: 'The requested content has been permanently deleted from server.'
        devMessage:
          type: string
          example: 'The requested content has been permanently deleted from server.'
        error:
          type: array
        data:
          type: array
    badEventResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 489
        type:
          type: string
          example: 'BAD_EVENT'
        message:
          type: string
          example: 'The server did not understand an event package specified in an Event header field.'
        devMessage:
          type: string
          example: 'The server did not understand an event package specified in an Event header field.'
        error:
          type: array
        data:
          type: array
    conflictResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 409
        type:
          type: string
          example: 'CONFLICT'
        message:
          type: string
          example: 'Resource already exist'
        devMessage:
          type: string
          example: 'Resource already exist'
        error:
          type: array
        data:
          type: array
    internalServerErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 500
        type:
          type: string
          example: 'Internal Server Error'
        message:
          type: string
          example: 'Internal Server Error'
        devMessage:
          type: string
          example: 'Internal Server Error'
        error:
          type: array
        data:
          type: array
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      summary: Register User
      description: An API to register new user in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/registerUserPayload'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/registerUserResponse'
        '400':
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/notFoundResponse'
        '409':
          description: CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/conflictResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalServerErrorResponse'
  /auth/verify-email/{userId}/{token}:
    put:
      operationId: verifyUser
      summary: Verify User Email ID
      description: An API to verify user email id in the system.
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/TokenParam'
      responses:
        '200':
          description: SUCCESS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/verifyUserResponse'
        '400':
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/notFoundResponse'
        '410':
          description: GONE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/goneResponse'
        '489':
          description: BAD_EVENT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badEventResponse'
        '500':
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalServerErrorResponse'
  /auth/login:
    post:
      operationId: loginUser
      summary: Login User
      description: An API to login a user which exist in system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginUserPayload'
      responses:
        '200':
          description: SUCCESS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/loginUserResponse'
        '400':
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/notFoundResponse'
        '410':
          description: GONE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/goneResponse'
        '500':
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalServerErrorResponse'
...